Restructure and refactoring

+++ internal/model/base.go
@@ -7,7 +7,7 @@ import (
 )
 
 type BaseEntity struct {
-	ID        int            `json:"id" gorm:"primaryKey"`
+	ID        int64          `json:"id" gorm:"primaryKey"`
 	CreatedAt time.Time      `json:"created_at"`
 	UpdatedAt time.Time      `json:"updated_at"`
 	DeletedAt gorm.DeletedAt `json:"-" gorm:"index"`
